Loads j/k orthogonal NIfTIs (still rudimentarily)
Makes the code able to load a slice while still receiving the NIfTI image.

The code does not check if the necessary bytes have already been received and processed.

No metadata has been provided yet.

Each pixel is occupying a lot of space and demanding lots of processing (matrix multiplication to find its x,y,z in patient coordinates). This was done because it is not possible anymore to use ndarray to change how to traverse the matrix (as we don't have the full matrix), and also to prepare for MPR (it seems Papaya holds the info this way too - need to check).